Enhancement: Extract LessThanTest
