+++ tests/Framework/ConstraintTest.php
@@ -1343,56 +1343,6 @@ EOF
         $this->fail();
     }
 
-    public function testConstraintLessThan()
-    {
-        $constraint = Assert::lessThan(1);
-
-        $this->assertTrue($constraint->evaluate(0, '', true));
-        $this->assertFalse($constraint->evaluate(1, '', true));
-        $this->assertEquals('is less than 1', $constraint->toString());
-        $this->assertCount(1, $constraint);
-
-        try {
-            $constraint->evaluate(1);
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-Failed asserting that 1 is less than 1.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
-    public function testConstraintLessThan2()
-    {
-        $constraint = Assert::lessThan(1);
-
-        try {
-            $constraint->evaluate(1, 'custom message');
-        } catch (ExpectationFailedException $e) {
-            $this->assertEquals(
-                <<<EOF
-custom message
-Failed asserting that 1 is less than 1.
-
-EOF
-                ,
-                TestFailure::exceptionToString($e)
-            );
-
-            return;
-        }
-
-        $this->fail();
-    }
-
     public function testConstraintNotLessThan()
     {
         $constraint = Assert::logicalNot(
